fix: 수익률을 float 타입으로 계산 시 소수점 오차가 발생하는 문제 해결

+++ src/main/java/lotto/domain/StatisticCalculator.java
@@ -23,12 +23,12 @@ public class StatisticCalculator {
         return repository;
     }
 
-    public float profitRate(int purchaseAmount, EnumMap<Rank, Integer> winningStatistics) {
+    public double profitRate(int purchaseAmount, EnumMap<Rank, Integer> winningStatistics) {
         long prize = winningStatistics.entrySet()
                 .stream()
                 .mapToLong(statistic -> (long) statistic.getKey().prizeMoney() * statistic.getValue())
                 .sum();
 
-        return (float) prize / purchaseAmount * 100;
+        return (double) prize / purchaseAmount * 100;
     }
 }
